pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "organising-tools"
version = "0.2.0"
description = "All-in-one organising tool CLI"
requires-python = ">=3.8"
dependencies = [
    "click",
    "rich",  # For pretty progress bars and logging
    "questionary", # For interactive menus
    "wakepy",  # Cross-platform sleep prevention
    "Pillow",  # For image compression
]

[project.scripts]
organising-tools = "organising_tools.cli:cli"
